added toggleable input mode. Separated fetch data code
